Close #361 - Rename Fx and FxCtor in sub-projects to fx and fxCtor respectively
