+++ cats-effect/src/main/scala-3/effectie/cats/Fx.scala
@@ -2,21 +2,21 @@ package effectie.cats
 
 import cats.effect.{IO, Sync}
 import cats.{Applicative, Id, Monad}
+import effectie.core.Fx
 
 import scala.concurrent.{ExecutionContext, Future}
 
-object Fx {
-  type Fx[F[*]] = effectie.core.Fx[F]
+object fx {
 
   given ioFx: Fx[IO] with {
 
-    inline override final def effectOf[A](a: => A): IO[A] = FxCtor.ioFxCtor.effectOf(a)
+    inline override final def effectOf[A](a: => A): IO[A] = fxCtor.ioFxCtor.effectOf(a)
 
-    inline override final def pureOf[A](a: A): IO[A] = FxCtor.ioFxCtor.pureOf(a)
+    inline override final def pureOf[A](a: A): IO[A] = fxCtor.ioFxCtor.pureOf(a)
 
-    inline override final def unitOf: IO[Unit] = FxCtor.ioFxCtor.unitOf
+    inline override final def unitOf: IO[Unit] = fxCtor.ioFxCtor.unitOf
 
-    inline override final def errorOf[A](throwable: Throwable): IO[A] = FxCtor.ioFxCtor.errorOf(throwable)
+    inline override final def errorOf[A](throwable: Throwable): IO[A] = fxCtor.ioFxCtor.errorOf(throwable)
 
     inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
@@ -29,23 +29,27 @@ object Fx {
     inline override final def handleNonFatal[A, AA >: A](fa: => IO[A])(handleError: Throwable => AA): IO[AA] =
       CanHandleError.ioCanHandleError.handleNonFatal(fa)(handleError)
 
-    inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => IO[A])(handleError: PartialFunction[Throwable, IO[AA]]): IO[AA] =
+    inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => IO[A])(
+      handleError: PartialFunction[Throwable, IO[AA]]
+    ): IO[AA] =
       CanRecover.ioCanRecover.recoverFromNonFatalWith(fa)(handleError)
 
-    inline override final def recoverFromNonFatal[A, AA >: A](fa: => IO[A])(handleError: PartialFunction[Throwable, AA]): IO[AA] =
+    inline override final def recoverFromNonFatal[A, AA >: A](fa: => IO[A])(
+      handleError: PartialFunction[Throwable, AA]
+    ): IO[AA] =
       CanRecover.ioCanRecover.recoverFromNonFatal(fa)(handleError)
 
   }
 
   given idFx: Fx[Id] with {
 
-    inline override final def effectOf[A](a: => A): Id[A] = FxCtor.idFxCtor.effectOf(a)
+    inline override final def effectOf[A](a: => A): Id[A] = fxCtor.idFxCtor.effectOf(a)
 
-    inline override final def pureOf[A](a: A): Id[A] = FxCtor.idFxCtor.pureOf(a)
+    inline override final def pureOf[A](a: A): Id[A] = fxCtor.idFxCtor.pureOf(a)
 
-    inline override final def unitOf: Id[Unit] = FxCtor.idFxCtor.unitOf
+    inline override final def unitOf: Id[Unit] = fxCtor.idFxCtor.unitOf
 
-    inline override final def errorOf[A](throwable: Throwable): Id[A] = FxCtor.idFxCtor.errorOf(throwable)
+    inline override final def errorOf[A](throwable: Throwable): Id[A] = fxCtor.idFxCtor.errorOf(throwable)
 
     inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
 
@@ -58,10 +62,14 @@ object Fx {
     inline override final def handleNonFatal[A, AA >: A](fa: => Id[A])(handleError: Throwable => AA): Id[AA] =
       CanHandleError.idCanHandleError.handleNonFatal(fa)(handleError)
 
-    inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => Id[A])(handleError: PartialFunction[Throwable, Id[AA]]): Id[AA] =
+    inline override final def recoverFromNonFatalWith[A, AA >: A](fa: => Id[A])(
+      handleError: PartialFunction[Throwable, Id[AA]]
+    ): Id[AA] =
       CanRecover.idCanRecover.recoverFromNonFatalWith(fa)(handleError)
 
-    inline override final def recoverFromNonFatal[A, AA >: A](fa: => Id[A])(handleError: PartialFunction[Throwable, AA]): Id[AA] =
+    inline override final def recoverFromNonFatal[A, AA >: A](fa: => Id[A])(
+      handleError: PartialFunction[Throwable, AA]
+    ): Id[AA] =
       CanRecover.idCanRecover.recoverFromNonFatal(fa)(handleError)
 
   }
