+++ cats-effect/src/test/scala-2/effectie/cats/CanHandleErrorSpec.scala
@@ -5,9 +5,10 @@ import cats.data.EitherT
 import cats.effect.IO
 import cats.instances.all._
 import cats.syntax.all._
-import effectie.cats.CanHandleError._
+import effectie.core._
 import effectie.syntax.fx._
-import effectie.cats.FxCtor._
+import effectie.cats.CanHandleError._
+import effectie.cats.fxCtor._
 import effectie.testing.types.SomeError
 import effectie.SomeControlThrowable
 import extras.concurrent.testing.ConcurrentSupport
@@ -22,9 +23,6 @@ import scala.util.control.{ControlThrowable, NonFatal}
   */
 object CanHandleErrorSpec extends Properties {
   implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
-
-  type FxCtor[F[_]] = effectie.core.FxCtor[F]
-  val FxCtor: effectie.core.FxCtor.type = effectie.core.FxCtor
 
   type CanHandleError[F[_]] = effectie.core.CanHandleError[F]
   val CanHandleError: effectie.core.CanHandleError.type = effectie.core.CanHandleError
