+++ cats-effect/src/test/scala-2/effectie/cats/fxSpec.scala
@@ -2,11 +2,10 @@ package effectie.cats
 
 import cats.data.EitherT
 import cats.effect._
-import cats.effect.unsafe.IORuntime
-import cats.syntax.either.catsSyntaxEitherId
+import cats.syntax.all._
 import cats.{Eq, Functor, Id, Monad}
-import effectie.cats.Fx._
-import effectie.cats.compat.CatsEffectIoCompatForFuture
+import effectie.core._
+import effectie.cats.fx._
 import effectie.testing.tools._
 import effectie.testing.types.{SomeError, SomeThrowableError}
 import effectie.SomeControlThrowable
@@ -15,17 +14,13 @@ import extras.concurrent.testing.types.{ErrorLogger, WaitFor}
 import hedgehog._
 import hedgehog.runner._
 
-import java.util.concurrent.ExecutorService
-import scala.concurrent.Await
 import scala.util.control.{ControlThrowable, NonFatal}
 
 /** @author Kevin Lee
   * @since 2020-12-06
   */
-object FxSpec extends Properties {
+object fxSpec extends Properties {
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
-
-  type Fx[F[_]] = effectie.core.Fx[F]
 
   override def tests: List[Test] = ioSpecs ++ futureSpecs ++ idSpecs
 
@@ -35,72 +30,64 @@ object FxSpec extends Properties {
     property("test Fx[IO].pureOf", IoSpec.testPureOf),
     example("test Fx[IO].unitOf", IoSpec.testUnitOf),
     example("test Fx[IO].errorOf", IoSpec.testErrorOf),
-    property("test Fx[IO] Monad laws - Identity", IoSpec.testMonadLaws1_Identity),
-    property("test Fx[IO] Monad laws - Composition", IoSpec.testMonadLaws2_Composition),
-    property("test Fx[IO] Monad laws - IdentityAp", IoSpec.testMonadLaws3_IdentityAp),
-    property("test Fx[IO] Monad laws - Homomorphism", IoSpec.testMonadLaws4_Homomorphism),
-    property("test Fx[IO] Monad laws - Interchange", IoSpec.testMonadLaws5_Interchange),
-    property("test Fx[IO] Monad laws - CompositionAp", IoSpec.testMonadLaws6_CompositionAp),
-    property("test Fx[IO] Monad laws - LeftIdentity", IoSpec.testMonadLaws7_LeftIdentity),
-    property("test Fx[IO] Monad laws - RightIdentity", IoSpec.testMonadLaws8_RightIdentity),
-    property("test Fx[IO] Monad laws - Associativity", IoSpec.testMonadLaws9_Associativity),
   ) ++
+    IoSpec.testMonadLaws ++
     List(
       example(
-        "test CanCatch[IO]catchNonFatalThrowable should catch NonFatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalThrowableShouldCatchNonFatal
+        "test Fx[IO]catchNonFatalThrowable should catch NonFatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalThrowableShouldCatchNonFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalThrowable should not catch Fatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalThrowableShouldNotCatchFatal
+        "test Fx[IO]catchNonFatalThrowable should not catch Fatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalThrowableShouldNotCatchFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalThrowable should return the successful result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalThrowableShouldReturnSuccessfulResult
+        "test Fx[IO]catchNonFatalThrowable should return the successful result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalThrowableShouldReturnSuccessfulResult
       ),
       example(
-        "test CanCatch[IO]catchNonFatal should catch NonFatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalShouldCatchNonFatal
+        "test Fx[IO]catchNonFatal should catch NonFatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalShouldCatchNonFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatal should not catch Fatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalShouldNotCatchFatal
+        "test Fx[IO]catchNonFatal should not catch Fatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalShouldNotCatchFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatal should return the successful result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalShouldReturnSuccessfulResult
+        "test Fx[IO]catchNonFatal should return the successful result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalShouldReturnSuccessfulResult
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEither should catch NonFatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherShouldCatchNonFatal
+        "test Fx[IO]catchNonFatalEither should catch NonFatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherShouldCatchNonFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEither should not catch Fatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherShouldNotCatchFatal
+        "test Fx[IO]catchNonFatalEither should not catch Fatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherShouldNotCatchFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEither should return the successful result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherShouldReturnSuccessfulResult
+        "test Fx[IO]catchNonFatalEither should return the successful result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherShouldReturnSuccessfulResult
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEither should return the failed result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherShouldReturnFailedResult
+        "test Fx[IO]catchNonFatalEither should return the failed result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherShouldReturnFailedResult
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEitherT should catch NonFatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherTShouldCatchNonFatal
+        "test Fx[IO]catchNonFatalEitherT should catch NonFatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherTShouldCatchNonFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEitherT should not catch Fatal",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherTShouldNotCatchFatal
+        "test Fx[IO]catchNonFatalEitherT should not catch Fatal",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherTShouldNotCatchFatal
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEitherT should return the successful result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherTShouldReturnSuccessfulResult
+        "test Fx[IO]catchNonFatalEitherT should return the successful result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherTShouldReturnSuccessfulResult
       ),
       example(
-        "test CanCatch[IO]catchNonFatalEitherT should return the failed result",
-        IoSpec.CanCatchSpec.testCanCatch_IO_catchNonFatalEitherTShouldReturnFailedResult
+        "test Fx[IO]catchNonFatalEitherT should return the failed result",
+        IoSpec.CanCatchSpec.testFx_IO_catchNonFatalEitherTShouldReturnFailedResult
       ),
     ) ++
     List(
@@ -345,7 +332,7 @@ object FxSpec extends Properties {
       example(
         "test Fx[IO].recoverEitherTFromNonFatal should return the failed result",
         IoSpec.CanRecoverSpec.testCanRecover_IO_recoverEitherTFromNonFatalShouldReturnFailedResult
-      ),
+      )
     )
 
   /* Future */
@@ -363,8 +350,7 @@ object FxSpec extends Properties {
         "test Fx[Future]catchNonFatalEitherT should return the failed result",
         FutureSpec.CanCatchSpec.testCanCatch_Future_catchNonFatalEitherTShouldReturnFailedResult
       ),
-    ) ++
-    List(
+    ) ++ List(
       example(
         "test Fx[Future].handleEitherTNonFatalWith should handle NonFatal",
         FutureSpec.CanHandleErrorSpec.testCanHandleError_Future_handleEitherTNonFatalWithShouldHandleNonFatalWith
@@ -413,7 +399,7 @@ object FxSpec extends Properties {
       example(
         "test Fx[Future].recoverEitherTFromNonFatal should return the failed result",
         FutureSpec.CanRecoverSpec.testCanRecover_Future_recoverEitherTFromNonFatalShouldReturnFailedResult
-      ),
+      )
     )
 
   /* Id */
@@ -422,17 +408,10 @@ object FxSpec extends Properties {
     property("test Fx[Id].pureOf", IdSpec.testPureOf),
     example("test Fx[Id].unitOf", IdSpec.testUnitOf),
     example("test Fx[Id].errorOf", IdSpec.testErrorOf),
-    property("test Fx[Id] Monad laws - Identity", IdSpec.testMonadLaws1_Identity),
-    property("test Fx[Id] Monad laws - Composition", IdSpec.testMonadLaws2_Composition),
-    property("test Fx[Id] Monad laws - IdentityAp", IdSpec.testMonadLaws3_IdentityAp),
-    property("test Fx[Id] Monad laws - Homomorphism", IdSpec.testMonadLaws4_Homomorphism),
-    property("test Fx[Id] Monad laws - Interchange", IdSpec.testMonadLaws5_Interchange),
-    property("test Fx[Id] Monad laws - CompositionAp", IdSpec.testMonadLaws6_CompositionAp),
-    property("test Fx[Id] Monad laws - LeftIdentity", IdSpec.testMonadLaws7_LeftIdentity),
-    property("test Fx[Id] Monad laws - RightIdentity", IdSpec.testMonadLaws8_RightIdentity),
-    property("test Fx[Id] Monad laws - Associativity", IdSpec.testMonadLaws9_Associativity),
   ) ++
+    IdSpec.testMonadLaws ++
     List(
+      /* Id */
       example(
         "test Fx[Id]catchNonFatalThrowable should catch NonFatal",
         IdSpec.CanCatchSpec.testCanCatch_Id_catchNonFatalThrowableShouldCatchNonFatal
@@ -612,6 +591,7 @@ object FxSpec extends Properties {
         IdSpec.CanHandleErrorSpec.testCanHandleError_Id_handleEitherTNonFatalShouldReturnFailedResult
       )
     ) ++ List(
+      /* Id */
       example(
         "test Fx[Id].recoverFromNonFatalWith should catch NonFatal",
         IdSpec.CanRecoverSpec.testCanRecover_Id_recoverFromNonFatalWithShouldRecoverFromNonFatal
@@ -734,7 +714,6 @@ object FxSpec extends Properties {
       )
     )
 
-  @SuppressWarnings(Array("org.wartremover.warts.Throw"))
   def throwThrowable[A](throwable: => Throwable): A =
     throw throwable
 
@@ -747,21 +726,15 @@ object FxSpec extends Properties {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
-      import CatsEffectRunner._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual        = before
       val testBefore    = actual ==== before
       val io            = Fx[IO].effectOf({ actual = after; () })
       val testBeforeRun = actual ==== before
-
-      val done         = io.completeAs(())
-      val testAfterRun = actual ==== after
-
+      io.unsafeRunSync()
+      val testAfterRun  = actual ==== after
       Result.all(
         List(
-          done,
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
           testAfterRun.log("testAfterRun")
@@ -773,20 +746,15 @@ object FxSpec extends Properties {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
       after  <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).map(_ + before).log("after")
     } yield {
-      import CatsEffectRunner._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
       @SuppressWarnings(Array("org.wartremover.warts.Var"))
       var actual        = before
       val testBefore    = actual ==== before
       val io            = Fx[IO].pureOf({ actual = after; () })
       val testBeforeRun = actual ==== after
-
-      val done         = io.completeAs(())
-      val testAfterRun = actual ==== after
+      io.unsafeRunSync()
+      val testAfterRun  = actual ==== after
       Result.all(
         List(
-          done,
           testBefore.log("testBefore"),
           testBeforeRun.log("testBeforeRun"),
           testAfterRun.log("testAfterRun")
@@ -795,145 +763,45 @@ object FxSpec extends Properties {
     }
 
     def testUnitOf: Result = {
-      import CatsEffectRunner._
-      implicit val ticket: Ticker = Ticker(TestContext())
-      val io                      = Fx[IO].unitOf
-      val expected: Unit          = ()
-      io.completeAs(expected)
+      val io             = Fx[IO].unitOf
+      val expected: Unit = ()
+      val actual: Unit   = io.unsafeRunSync()
+      actual ==== expected
     }
 
     def testErrorOf: Result = {
       val expectedMessage = "This is a throwable test error."
       val expectedError   = SomeThrowableError.message(expectedMessage)
 
-      import CatsEffectRunner._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      val io = Fx[IO].errorOf[Unit](expectedError)
-      io.expectError(expectedError)
+      val io = Fx[IO].errorOf(expectedError)
+      expectThrowable(io.unsafeRunSync(), expectedError)
     }
 
-    def testMonadLaws1_Identity: Property = {
-      import CatsEffectRunner._
+    def testMonadLaws: List[Test] = {
       import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
 
       implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
+        (x, y) => x.flatMap(xx => y.map(_ === xx)).unsafeRunSync()
 
-      MonadSpec.test1_Identity[IO]
-    }
+      implicit val ioFx: Fx[IO] = effectie.cats.fx.IoFx
 
-    def testMonadLaws2_Composition: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test2_Composition[IO]
-    }
-
-    def testMonadLaws3_IdentityAp: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test3_IdentityAp[IO]
-    }
-
-    def testMonadLaws4_Homomorphism: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test4_Homomorphism[IO]
-    }
-
-    def testMonadLaws5_Interchange: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test5_Interchange[IO]
-    }
-
-    def testMonadLaws6_CompositionAp: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test6_CompositionAp[IO]
-    }
-
-    def testMonadLaws7_LeftIdentity: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test7_LeftIdentity[IO]
-    }
-
-    def testMonadLaws8_RightIdentity: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-      MonadSpec.test8_RightIdentity[IO]
-    }
-
-    def testMonadLaws9_Associativity: Property = {
-      import CatsEffectRunner._
-      import cats.syntax.eq._
-      implicit val ticket: Ticker = Ticker(TestContext())
-
-      implicit val eqIo: Eq[IO[Int]] =
-        (x, y) => x.flatMap(xx => y.map(_ === xx)).completeAndEqualTo(true)
-
-//      implicit val ioFx: Fx[IO] = Fx.IoFx
-
-      MonadSpec.test9_Associativity[IO]
+      MonadSpec.testMonadLaws[IO]("IO")
     }
 
     object CanCatchSpec {
 
-      def testCanCatch_IO_catchNonFatalThrowableShouldCatchNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalThrowableShouldCatchNonFatal: Result = {
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
         val expected          = expectedExpcetion.asLeft[Int]
-        val actual            = Fx[IO].catchNonFatalThrowable(fa)
+        val actual            = Fx[IO].catchNonFatalThrowable(fa).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-      def testCanCatch_IO_catchNonFatalThrowableShouldNotCatchFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
+      def testFx_IO_catchNonFatalThrowableShouldNotCatchFatal: Result = {
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
@@ -951,36 +819,27 @@ object FxSpec extends Properties {
 
       }
 
-      def testCanCatch_IO_catchNonFatalThrowableShouldReturnSuccessfulResult: Result = {
+      def testFx_IO_catchNonFatalThrowableShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+        val fa       = run[IO, Int](1)
+        val expected = 1.asRight[Throwable]
+        val actual   = Fx[IO].catchNonFatalThrowable(fa).unsafeRunSync()
 
-        val fa: IO[Int] = run[IO, Int](1)
-        val expected    = 1.asRight[Throwable]
-        val actual      = Fx[IO].catchNonFatalThrowable(fa)
-
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
-      def testCanCatch_IO_catchNonFatalShouldCatchNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalShouldCatchNonFatal: Result = {
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
         val expected          = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actual            = Fx[IO].catchNonFatal(fa)(SomeError.someThrowable)
+        val actual            = Fx[IO].catchNonFatal(fa)(SomeError.someThrowable).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-      def testCanCatch_IO_catchNonFatalShouldNotCatchFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
+      def testFx_IO_catchNonFatalShouldNotCatchFatal: Result = {
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
@@ -998,36 +857,27 @@ object FxSpec extends Properties {
 
       }
 
-      def testCanCatch_IO_catchNonFatalShouldReturnSuccessfulResult: Result = {
+      def testFx_IO_catchNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+        val fa       = run[IO, Int](1)
+        val expected = 1.asRight[SomeError]
+        val actual   = Fx[IO].catchNonFatal(fa)(SomeError.someThrowable).unsafeRunSync()
 
-        val fa: IO[Int] = run[IO, Int](1)
-        val expected    = 1.asRight[SomeError]
-        val actual      = Fx[IO].catchNonFatal(fa)(SomeError.someThrowable)
-
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
-      def testCanCatch_IO_catchNonFatalEitherShouldCatchNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalEitherShouldCatchNonFatal: Result = {
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa       = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
         val expected = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actual   = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable)
+        val actual   = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-      def testCanCatch_IO_catchNonFatalEitherShouldNotCatchFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
+      def testFx_IO_catchNonFatalEitherShouldNotCatchFatal: Result = {
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
@@ -1045,49 +895,39 @@ object FxSpec extends Properties {
 
       }
 
-      def testCanCatch_IO_catchNonFatalEitherShouldReturnSuccessfulResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalEitherShouldReturnSuccessfulResult: Result = {
 
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable)
+        val actual   = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
-      def testCanCatch_IO_catchNonFatalEitherShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalEitherShouldReturnFailedResult: Result = {
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
-        val actual          = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable)
+        val actual          = Fx[IO].catchNonFatalEither(fa)(SomeError.someThrowable).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
-      def testCanCatch_IO_catchNonFatalEitherTShouldCatchNonFatal: Result = {
+      def testFx_IO_catchNonFatalEitherTShouldCatchNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+        val expectedExpcetion               = new RuntimeException("Something's wrong")
+        val fa: EitherT[IO, SomeError, Int] = EitherT(
+          run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
+        )
+        val expected                        = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actual = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
 
-        val expectedExpcetion = new RuntimeException("Something's wrong")
-        val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-        val expected = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actual   = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value
-
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
-      def testCanCatch_IO_catchNonFatalEitherTShouldNotCatchFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
+      def testFx_IO_catchNonFatalEitherTShouldNotCatchFatal: Result = {
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion)))
@@ -1105,29 +945,23 @@ object FxSpec extends Properties {
 
       }
 
-      def testCanCatch_IO_catchNonFatalEitherTShouldReturnSuccessfulResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalEitherTShouldReturnSuccessfulResult: Result = {
 
         val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value
+        val actual   = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
-      def testCanCatch_IO_catchNonFatalEitherTShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
+      def testFx_IO_catchNonFatalEitherTShouldReturnFailedResult: Result = {
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
         val expected        = expectedFailure.asLeft[Int]
-        val actual          = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value
+        val actual          = Fx[IO].catchNonFatalEitherT(fa)(SomeError.someThrowable).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
     }
@@ -1135,9 +969,6 @@ object FxSpec extends Properties {
     object CanHandleErrorSpec {
 
       def testCanHandleError_IO_handleNonFatalWithShouldHandleNonFatalWith: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
@@ -1147,15 +978,13 @@ object FxSpec extends Properties {
             case NonFatal(`expectedExpcetion`) =>
               IO.pure(expected)
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleNonFatalWithShouldNotHandleFatalWith: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
@@ -1175,39 +1004,30 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Int](1)
         val expected = 1
-        val actual   = Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(999))
+        val actual   = Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(999)).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleNonFatalWithEitherShouldHandleNonFatalWith: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
         val expectedFailedResult = SomeError.message("Recovered Error").asLeft[Int]
         val actualFailedResult   =
-          Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(expectedFailedResult))
+          Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(expectedFailedResult)).unsafeRunSync()
 
         val expectedSuccessResult = 1.asRight[SomeError]
         val actualSuccessResult   =
-          Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(1.asRight[SomeError]))
+          Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(1.asRight[SomeError])).unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleNonFatalWithEitherShouldNotHandleFatalWith: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
@@ -1227,52 +1047,40 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleNonFatalWithEitherShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].handleNonFatalWith(fa)(_ => IO(999.asRight[SomeError]))
+        val actual   = Fx[IO].handleNonFatalWith(fa)(_ => IO(999.asRight[SomeError])).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleNonFatalWithEitherShouldReturnFailedResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
-        val actual          = Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError]))
+        val actual          = Fx[IO].handleNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherNonFatalWithShouldHandleNonFatalWith: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .handleEitherNonFatalWith(fa)(err => IO.pure(SomeError.someThrowable(err).asLeft[Int]))
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
-          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError]))
+          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleEitherNonFatalWithShouldNotHandleFatalWith: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
@@ -1295,55 +1103,43 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleEitherNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
         val actual   =
-          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError]))
+          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherNonFatalWithShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
         val actual          =
-          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError]))
+          Fx[IO].handleEitherNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherTNonFatalWithShouldHandleNonFatalWith: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .handleEitherTNonFatalWith(fa)(err => IO.pure(SomeError.someThrowable(err).asLeft[Int]))
           .value
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
-          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value
+          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value.unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleEitherTNonFatalWithShouldNotHandleFatalWith: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion)))
@@ -1367,35 +1163,26 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleEitherTNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
         val expected = 1.asRight[SomeError]
         val actual   =
-          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value
+          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherTNonFatalWithShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
         val expected        = expectedFailure.asLeft[Int]
         val actual          =
-          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value
+          Fx[IO].handleEitherTNonFatalWith(fa)(_ => IO.pure(123.asRight[SomeError])).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleNonFatalShouldHandleNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
@@ -1405,15 +1192,13 @@ object FxSpec extends Properties {
             case NonFatal(`expectedExpcetion`) =>
               expected
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleNonFatalShouldNotHandleFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Int](throwThrowable[Int](fatalExpcetion))
@@ -1433,37 +1218,28 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Int](1)
         val expected = 1
-        val actual   = Fx[IO].handleNonFatal(fa)(_ => 999)
+        val actual   = Fx[IO].handleNonFatal(fa)(_ => 999).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleNonFatalEitherShouldHandleNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
         val expectedFailedResult = SomeError.message("Recovered Error").asLeft[Int]
-        val actualFailedResult   = Fx[IO].handleNonFatal(fa)(_ => expectedFailedResult)
+        val actualFailedResult   = Fx[IO].handleNonFatal(fa)(_ => expectedFailedResult).unsafeRunSync()
 
         val expectedSuccessResult = 1.asRight[SomeError]
-        val actualSuccessResult   = Fx[IO].handleNonFatal(fa)(_ => 1.asRight[SomeError])
+        val actualSuccessResult   = Fx[IO].handleNonFatal(fa)(_ => 1.asRight[SomeError]).unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleNonFatalEitherShouldNotHandleFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
@@ -1483,52 +1259,40 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleNonFatalEitherShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].handleNonFatal(fa)(_ => 999.asRight[SomeError])
+        val actual   = Fx[IO].handleNonFatal(fa)(_ => 999.asRight[SomeError]).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleNonFatalEitherShouldReturnFailedResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
-        val actual          = Fx[IO].handleNonFatal(fa)(_ => 123.asRight[SomeError])
+        val actual          = Fx[IO].handleNonFatal(fa)(_ => 123.asRight[SomeError]).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherNonFatalShouldHandleNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .handleEitherNonFatal(fa)(err => SomeError.someThrowable(err).asLeft[Int])
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
-          Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError])
+          Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError]).unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleEitherNonFatalShouldNotHandleFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa             = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion))
@@ -1551,54 +1315,42 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleEitherNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError])
+        val actual   = Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError]).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherNonFatalShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
         val actual          =
-          Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError])
+          Fx[IO].handleEitherNonFatal(fa)(_ => 123.asRight[SomeError]).unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherTNonFatalShouldHandleNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .handleEitherTNonFatal(fa)(err => SomeError.someThrowable(err).asLeft[Int])
           .value
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
-          Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value
+          Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value.unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanHandleError_IO_handleEitherTNonFatalShouldNotHandleFatal: Result = {
-
-        val es: ExecutorService    = ConcurrentSupport.newExecutorService(2)
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(es)
 
         val fatalExpcetion = SomeControlThrowable("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](fatalExpcetion)))
@@ -1622,28 +1374,22 @@ object FxSpec extends Properties {
 
       def testCanHandleError_IO_handleEitherTNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value
+        val actual   = Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanHandleError_IO_handleEitherTNonFatalShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
         val expected        = expectedFailure.asLeft[Int]
         val actual          =
-          Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value
+          Fx[IO].handleEitherTNonFatal(fa)(_ => 123.asRight[SomeError]).value.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
     }
@@ -1652,25 +1398,21 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverFromNonFatalWithShouldRecoverFromNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
         val expected          = 123
-        val actual            = Fx[IO].recoverFromNonFatalWith(fa) {
-          case NonFatal(`expectedExpcetion`) =>
-            IO.pure(expected)
-        }
-        actual.completeAs(expected)
+        val actual            = Fx[IO]
+          .recoverFromNonFatalWith(fa) {
+            case NonFatal(`expectedExpcetion`) =>
+              IO.pure(expected)
+          }
+          .unsafeRunSync()
 
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverFromNonFatalWithShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
@@ -1691,22 +1433,18 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverFromNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
+        val fa       = run[IO, Int](1)
         val expected = 1
-        val fa       = run[IO, Int](expected)
         val actual   = Fx[IO]
           .recoverFromNonFatalWith(fa) {
             case NonFatal(_) => IO.pure(999)
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverFromNonFatalWithEitherShouldRecoverFromNonFatal: Result = {
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
@@ -1715,21 +1453,20 @@ object FxSpec extends Properties {
           .recoverFromNonFatalWith(fa) {
             case NonFatal(`expectedExpcetion`) => IO.pure(expectedFailedResult)
           }
+          .unsafeRunSync()
 
         val expectedSuccessResult = 1.asRight[SomeError]
         val actualSuccessResult   = Fx[IO]
           .recoverFromNonFatalWith(fa) {
             case NonFatal(`expectedExpcetion`) => IO.pure(1.asRight[SomeError])
           }
+          .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverFromNonFatalWithEitherShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
@@ -1752,23 +1489,18 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverFromNonFatalWithEitherShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
         val actual   = Fx[IO]
           .recoverFromNonFatalWith(fa) {
             case NonFatal(_) => IO(999.asRight[SomeError])
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverFromNonFatalWithEitherShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
@@ -1777,37 +1509,33 @@ object FxSpec extends Properties {
           .recoverFromNonFatalWith(fa) {
             case NonFatal(_) => IO.pure(123.asRight[SomeError])
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherFromNonFatalWithShouldRecoverFromNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .recoverEitherFromNonFatalWith(fa) {
             case err => IO.pure(SomeError.someThrowable(err).asLeft[Int])
           }
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   = Fx[IO]
           .recoverEitherFromNonFatalWith(fa) {
             case NonFatal(`expectedExpcetion`) => IO.pure(123.asRight[SomeError])
           }
+          .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverEitherFromNonFatalWithShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
@@ -1830,23 +1558,18 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverEitherFromNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
         val actual   = Fx[IO]
           .recoverEitherFromNonFatalWith(fa) {
             case NonFatal(_) => IO.pure(123.asRight[SomeError])
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherFromNonFatalWithShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
@@ -1856,39 +1579,35 @@ object FxSpec extends Properties {
             .recoverEitherFromNonFatalWith(fa) {
               case NonFatal(_) => IO.pure(123.asRight[SomeError])
             }
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherTFromNonFatalWithShouldRecoverFromNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   = Fx[IO]
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    = Fx[IO]
           .recoverEitherTFromNonFatalWith(fa) {
             case err => IO.pure(SomeError.someThrowable(err).asLeft[Int])
           }
           .value
-
+          .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   = Fx[IO]
           .recoverEitherTFromNonFatalWith(fa) {
             case NonFatal(`expectedExpcetion`) => IO.pure(123.asRight[SomeError])
           }
           .value
+          .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverEitherTFromNonFatalWithShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
@@ -1911,9 +1630,6 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverEitherTFromNonFatalWithShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
         val expected = 1.asRight[SomeError]
         val actual   = Fx[IO]
@@ -1921,14 +1637,12 @@ object FxSpec extends Properties {
             case NonFatal(_) => IO.pure(123.asRight[SomeError])
           }
           .value
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherTFromNonFatalWithShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
@@ -1939,16 +1653,14 @@ object FxSpec extends Properties {
               case NonFatal(_) => IO.pure(123.asRight[SomeError])
             }
             .value
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       // /
 
       def testCanRecover_IO_recoverFromNonFatalShouldRecoverFromNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
@@ -1958,15 +1670,13 @@ object FxSpec extends Properties {
             case NonFatal(`expectedExpcetion`) =>
               expected
           }
+          .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverFromNonFatalShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa                = run[IO, Int](throwThrowable[Int](expectedExpcetion))
@@ -1987,39 +1697,32 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverFromNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Int](1)
         val expected = 1
-        val actual   = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 999 }
+        val actual   = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 999 }.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverFromNonFatalEitherShouldRecoverFromNonFatal: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
         val expectedFailedResult = SomeError.message("Recovered Error").asLeft[Int]
         val actualFailedResult   = Fx[IO]
           .recoverFromNonFatal(fa) { case NonFatal(`expectedExpcetion`) => expectedFailedResult }
+          .unsafeRunSync()
 
         val expectedSuccessResult = 1.asRight[SomeError]
         val actualSuccessResult   = Fx[IO]
           .recoverFromNonFatal(fa) { case NonFatal(`expectedExpcetion`) => 1.asRight[SomeError] }
+          .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverFromNonFatalEitherShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
@@ -2040,56 +1743,45 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverFromNonFatalEitherShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
-        val actual   = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 999.asRight[SomeError] }
+        val actual   = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 999.asRight[SomeError] }.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverFromNonFatalEitherShouldReturnFailedResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
         val expected        = expectedFailure.asLeft[Int]
-        val actual          = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }
+        val actual = Fx[IO].recoverFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }.unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherFromNonFatalShouldRecoverFromNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   =
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    =
           Fx[IO]
             .recoverEitherFromNonFatal(fa) {
               case err => SomeError.someThrowable(err).asLeft[Int]
             }
-
+            .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
           Fx[IO]
             .recoverEitherFromNonFatal(fa) { case NonFatal(`expectedExpcetion`) => 123.asRight[SomeError] }
+            .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverEitherFromNonFatalShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
@@ -2113,22 +1805,17 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverEitherFromNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = run[IO, Either[SomeError, Int]](1.asRight[SomeError])
         val expected = 1.asRight[SomeError]
         val actual   =
           Fx[IO]
             .recoverEitherFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherFromNonFatalShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int])
@@ -2136,39 +1823,35 @@ object FxSpec extends Properties {
         val actual          =
           Fx[IO]
             .recoverEitherFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherTFromNonFatalShouldRecoverFromNonFatal: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val expectedExpcetion = new RuntimeException("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
-        val expectedFailedResult = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
-        val actualFailedResult   =
+        val expectedFailedResult  = SomeError.someThrowable(expectedExpcetion).asLeft[Int]
+        val actualFailedResult    =
           Fx[IO]
             .recoverEitherTFromNonFatal(fa) {
               case err => SomeError.someThrowable(err).asLeft[Int]
             }
             .value
-
+            .unsafeRunSync()
         val expectedSuccessResult = 123.asRight[SomeError]
         val actualSuccessResult   =
           Fx[IO]
             .recoverEitherTFromNonFatal(fa) { case NonFatal(`expectedExpcetion`) => 123.asRight[SomeError] }
             .value
+            .unsafeRunSync()
 
-        actualFailedResult.completeAs(expectedFailedResult) and actualSuccessResult.completeAs(expectedSuccessResult)
+        actualFailedResult ==== expectedFailedResult and actualSuccessResult ==== expectedSuccessResult
       }
 
       @SuppressWarnings(Array("org.wartremover.warts.ToString"))
       def testCanRecover_IO_recoverEitherTFromNonFatalShouldNotCatchFatal: Result = {
-
-        val compat                 = new CatsEffectIoCompatForFuture
-        implicit val rt: IORuntime = testing.IoAppUtils.runtime(compat.es)
 
         val expectedExpcetion = SomeControlThrowable("Something's wrong")
         val fa = EitherT(run[IO, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion)))
@@ -2192,23 +1875,18 @@ object FxSpec extends Properties {
 
       def testCanRecover_IO_recoverEitherTFromNonFatalShouldReturnSuccessfulResult: Result = {
 
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
-
         val fa       = EitherT(run[IO, Either[SomeError, Int]](1.asRight[SomeError]))
         val expected = 1.asRight[SomeError]
         val actual   =
           Fx[IO]
             .recoverEitherTFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }
             .value
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
       def testCanRecover_IO_recoverEitherTFromNonFatalShouldReturnFailedResult: Result = {
-
-        import CatsEffectRunner._
-        implicit val ticket: Ticker = Ticker(TestContext())
 
         val expectedFailure = SomeError.message("Failed")
         val fa              = EitherT(run[IO, Either[SomeError, Int]](expectedFailure.asLeft[Int]))
@@ -2217,26 +1895,23 @@ object FxSpec extends Properties {
           Fx[IO]
             .recoverEitherTFromNonFatal(fa) { case NonFatal(_) => 123.asRight[SomeError] }
             .value
+            .unsafeRunSync()
 
-        actual.completeAs(expected)
+        actual ==== expected
       }
 
     }
+
   }
 
   object FutureSpec {
-    import java.util.concurrent.{ExecutorService, Executors}
     import scala.concurrent.duration._
-    import scala.concurrent.{ExecutionContext, Future}
 
     val waitFor = WaitFor(1.second)
 
-    implicit def futureEqual[A: Eq](
-      implicit ec: scala.concurrent.ExecutionContext
-    ): Eq[Future[A]] =
-      (x: Future[A], y: Future[A]) => Await.result(x.flatMap(a => y.map(b => Eq[A].eqv(a, b))), waitFor.waitFor)
-
     object CanCatchSpec {
+      import java.util.concurrent.{ExecutorService, Executors}
+      import scala.concurrent.{ExecutionContext, Future}
 
       def testCanCatch_Future_catchNonFatalEitherTShouldCatchNonFatal: Result = {
 
@@ -2288,6 +1963,7 @@ object FxSpec extends Properties {
         actual ==== expected
       }
     }
+
     object CanHandleErrorSpec {
       import java.util.concurrent.{ExecutorService, Executors}
       import scala.concurrent.{ExecutionContext, Future}
@@ -2445,11 +2121,12 @@ object FxSpec extends Properties {
           waitFor
         )
 
-        val fa2      = EitherT(
+        val expected = 1.asRight[SomeError]
+
+        val fa2    = EitherT(
           run[Future, Either[SomeError, Int]](throwThrowable[Either[SomeError, Int]](expectedExpcetion))
         )
-        val expected = 1.asRight[SomeError]
-        val actual   =
+        val actual =
           ConcurrentSupport.futureToValueAndTerminate(
             executorService,
             waitFor
@@ -2631,34 +2308,10 @@ object FxSpec extends Properties {
       expectThrowable(actual, expectedError)
     }
 
-    implicit val idInstance: Monad[Id] = cats.catsInstancesForId
-
-    def testMonadLaws1_Identity: Property =
-      MonadSpec.test1_Identity[Id]
-
-    def testMonadLaws2_Composition: Property =
-      MonadSpec.test2_Composition[Id]
-
-    def testMonadLaws3_IdentityAp: Property =
-      MonadSpec.test3_IdentityAp[Id]
-
-    def testMonadLaws4_Homomorphism: Property =
-      MonadSpec.test4_Homomorphism[Id]
-
-    def testMonadLaws5_Interchange: Property =
-      MonadSpec.test5_Interchange[Id]
-
-    def testMonadLaws6_CompositionAp: Property =
-      MonadSpec.test6_CompositionAp[Id]
-
-    def testMonadLaws7_LeftIdentity: Property =
-      MonadSpec.test7_LeftIdentity[Id]
-
-    def testMonadLaws8_RightIdentity: Property =
-      MonadSpec.test8_RightIdentity[Id]
-
-    def testMonadLaws9_Associativity: Property =
-      MonadSpec.test9_Associativity[Id]
+    def testMonadLaws: List[Test] = {
+      val idInstance: Monad[Id] = cats.catsInstancesForId
+      MonadSpec.testMonadLaws[Id]("Id")
+    }
 
     object CanCatchSpec {
 
@@ -3600,6 +3253,7 @@ object FxSpec extends Properties {
         try {
           val actual = Fx[Id]
             .recoverEitherFromNonFatal(fa) { case NonFatal(`expectedExpcetion`) => 1.asRight[SomeError] }
+
           Result.failure.log(s"The expected fatal exception was not thrown. actual: ${actual.toString}")
         } catch {
           case ex: ControlThrowable =>
