+++ cats-effect/src/test/scala-2/effectie/cats/syntax/errorSpec.scala
@@ -4,7 +4,7 @@ import cats._
 import cats.data.EitherT
 import cats.effect._
 import cats.syntax.all._
-import effectie.cats.Fx._
+import effectie.cats.fx._
 import effectie.cats.syntax.error._
 import effectie.syntax.error._
 import effectie.syntax.fx._
