+++ cats-effect/src/test/scala-3/effectie/cats/CanCatchSpec.scala
@@ -6,7 +6,7 @@ import cats.effect.*
 import cats.instances.all.*
 import cats.syntax.all.*
 import effectie.cats.CanCatch as *
-import effectie.cats.Fx.given
+import effectie.cats.fx.given
 import effectie.syntax.fx.*
 import effectie.testing.types.SomeError
 import effectie.core.{CanCatch, Fx}
