+++ cats-effect/src/test/scala-3/effectie/cats/FxCtorSpec.scala
@@ -34,7 +34,7 @@ object FxCtorSpec extends Properties {
 
   object IoSpec {
 
-    import effectie.cats.FxCtor.given
+    import effectie.cats.fxCtor.given
 
     def testEffectOf: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
@@ -92,7 +92,7 @@ object FxCtorSpec extends Properties {
 
   object IdSpec {
 
-    import effectie.cats.FxCtor.given
+    import effectie.cats.fxCtor.given
 
     def testEffectOf: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
