+++ cats-effect3/src/main/scala-2/effectie/cats/Catching.scala
@@ -1,5 +1,6 @@
 package effectie.cats
 
+import effectie.core._
 import cats.data.EitherT
 
 /** @author Kevin Lee
