+++ cats-effect3/src/main/scala-2/effectie/cats/Fx.scala
@@ -3,20 +3,18 @@ package effectie.cats
 import cats.effect.IO
 import cats.Id
 
-object Fx {
-  type Fx[F[_]] = effectie.core.Fx[F]
-
-  def apply[F[_]: Fx]: Fx[F] = implicitly[Fx[F]]
+object fx {
+  private type Fx[F[_]] = effectie.core.Fx[F]
 
   implicit object IoFx extends Fx[IO] {
 
-    @inline override final def effectOf[A](a: => A): IO[A] = FxCtor.IoFxCtor.effectOf(a)
+    @inline override final def effectOf[A](a: => A): IO[A] = fxCtor.IoFxCtor.effectOf(a)
 
-    @inline override final def pureOf[A](a: A): IO[A] = FxCtor.IoFxCtor.pureOf(a)
+    @inline override final def pureOf[A](a: A): IO[A] = fxCtor.IoFxCtor.pureOf(a)
 
-    @inline override final val unitOf: IO[Unit] = FxCtor.IoFxCtor.unitOf
+    @inline override final val unitOf: IO[Unit] = fxCtor.IoFxCtor.unitOf
 
-    @inline override final def errorOf[A](throwable: Throwable): IO[A] = FxCtor.IoFxCtor.errorOf(throwable)
+    @inline override final def errorOf[A](throwable: Throwable): IO[A] = fxCtor.IoFxCtor.errorOf(throwable)
 
     @inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
@@ -42,13 +40,13 @@ object Fx {
 
   implicit object IdFx extends Fx[Id] {
 
-    @inline override final def effectOf[A](a: => A): Id[A] = FxCtor.IdFxCtor.effectOf(a)
+    @inline override final def effectOf[A](a: => A): Id[A] = fxCtor.IdFxCtor.effectOf(a)
 
-    @inline override final def pureOf[A](a: A): Id[A] = FxCtor.IdFxCtor.pureOf(a)
+    @inline override final def pureOf[A](a: A): Id[A] = fxCtor.IdFxCtor.pureOf(a)
 
-    @inline override final val unitOf: Id[Unit] = FxCtor.IdFxCtor.unitOf
+    @inline override final val unitOf: Id[Unit] = fxCtor.IdFxCtor.unitOf
 
-    @inline override final def errorOf[A](throwable: Throwable): Id[A] = FxCtor.IdFxCtor.errorOf(throwable)
+    @inline override final def errorOf[A](throwable: Throwable): Id[A] = fxCtor.IdFxCtor.errorOf(throwable)
 
     @inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
