+++ cats-effect3/src/main/scala-2/effectie/cats/FxCtor.scala
@@ -3,10 +3,8 @@ package effectie.cats
 import cats.Id
 import cats.effect.IO
 
-object FxCtor {
-  type FxCtor[F[_]] = effectie.core.FxCtor[F]
-
-  def apply[F[_]: FxCtor]: FxCtor[F] = implicitly[FxCtor[F]]
+object fxCtor {
+  private type FxCtor[F[_]] = effectie.core.FxCtor[F]
 
   implicit object IoFxCtor extends FxCtor[IO] {
