+++ cats-effect3/src/test/scala-2/effectie/cats/CanRecoverSpec.scala
@@ -7,7 +7,7 @@ import cats.effect.unsafe.IORuntime
 import cats.instances.all._
 import cats.syntax.all._
 import effectie.cats.CanRecover._
-import effectie.cats.FxCtor._
+import effectie.cats.fxCtor._
 import effectie.syntax.fx._
 import effectie.cats.compat.CatsEffectIoCompatForFuture
 import effectie.testing.types.SomeError
