+++ cats-effect3/src/test/scala-2/effectie/cats/fxCtorSpec.scala
@@ -2,6 +2,8 @@ package effectie.cats
 
 import cats.Id
 import cats.effect._
+import effectie.core._
+import effectie.cats.fxCtor._
 import effectie.testing.tools
 import effectie.testing.types.SomeThrowableError
 import extras.concurrent.testing.types.ErrorLogger
@@ -11,7 +13,7 @@ import hedgehog.runner._
 /** @author Kevin Lee
   * @since 2020-12-06
   */
-object FxCtorSpec extends Properties {
+object fxCtorSpec extends Properties {
   private implicit val errorLogger: ErrorLogger[Throwable] = ErrorLogger.printlnDefaultErrorLogger
 
   override def tests: List[Test] = ioSpecs ++ futureSpecs ++ idSpecs
@@ -33,7 +35,6 @@ object FxCtorSpec extends Properties {
   )
 
   object IoSpec {
-    import effectie.cats.Fx._
 
     def testEffectOf: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
@@ -107,8 +108,6 @@ object FxCtorSpec extends Properties {
   }
 
   object IdSpec {
-
-    import effectie.cats.Fx._
 
     def testEffectOf: Property = for {
       before <- Gen.int(Range.linear(Int.MinValue, Int.MaxValue)).log("before")
