+++ effectie-monix/src/main/scala/effectie/monix/CanCatch.scala
@@ -9,7 +9,7 @@ import monix.eval.Task
   * @since 2020-06-07
   */
 object CanCatch {
-  type CanCatch[F[_]] = effectie.core.CanCatch[F]
+  private type CanCatch[F[_]] = effectie.core.CanCatch[F]
 
   implicit object CanCatchTask extends CanCatch[Task] {
