+++ effectie-monix/src/main/scala/effectie/monix/Fx.scala
@@ -7,18 +7,18 @@ import monix.eval.Task
 /** @author Kevin Lee
   * @since 2021-05-16
   */
-object Fx {
-  type Fx[F[_]] = effectie.core.Fx[F]
+object fx {
+  private type Fx[F[_]] = effectie.core.Fx[F]
 
   implicit object TaskFx extends Fx[Task] {
 
-    @inline override final def effectOf[A](a: => A): Task[A] = FxCtor.TaskFxCtor.effectOf(a)
+    @inline override final def effectOf[A](a: => A): Task[A] = fxCtor.TaskFxCtor.effectOf(a)
 
-    @inline override final def pureOf[A](a: A): Task[A] = FxCtor.TaskFxCtor.pureOf(a)
+    @inline override final def pureOf[A](a: A): Task[A] = fxCtor.TaskFxCtor.pureOf(a)
 
-    @inline override final val unitOf: Task[Unit] = FxCtor.TaskFxCtor.unitOf
+    @inline override final val unitOf: Task[Unit] = fxCtor.TaskFxCtor.unitOf
 
-    @inline override final def errorOf[A](throwable: Throwable): Task[A] = FxCtor.TaskFxCtor.errorOf(throwable)
+    @inline override final def errorOf[A](throwable: Throwable): Task[A] = fxCtor.TaskFxCtor.errorOf(throwable)
 
     @inline override final def mapFa[A, B](fa: Task[A])(f: A => B): Task[B] = fa.map(f)
 
@@ -46,13 +46,13 @@ object Fx {
 
   implicit object IoFx extends Fx[IO] {
 
-    @inline override final def effectOf[A](a: => A): IO[A] = FxCtor.IoFxCtor.effectOf(a)
+    @inline override final def effectOf[A](a: => A): IO[A] = fxCtor.IoFxCtor.effectOf(a)
 
-    @inline override final def pureOf[A](a: A): IO[A] = FxCtor.IoFxCtor.pureOf(a)
+    @inline override final def pureOf[A](a: A): IO[A] = fxCtor.IoFxCtor.pureOf(a)
 
-    @inline override final val unitOf: IO[Unit] = FxCtor.IoFxCtor.unitOf
+    @inline override final val unitOf: IO[Unit] = fxCtor.IoFxCtor.unitOf
 
-    @inline override final def errorOf[A](throwable: Throwable): IO[A] = FxCtor.IoFxCtor.errorOf(throwable)
+    @inline override final def errorOf[A](throwable: Throwable): IO[A] = fxCtor.IoFxCtor.errorOf(throwable)
 
     @inline override final def mapFa[A, B](fa: IO[A])(f: A => B): IO[B] = fa.map(f)
 
@@ -78,13 +78,13 @@ object Fx {
 
   implicit object IdFx extends Fx[Id] {
 
-    @inline override final def effectOf[A](a: => A): Id[A] = FxCtor.IdFxCtor.effectOf(a)
+    @inline override final def effectOf[A](a: => A): Id[A] = fxCtor.IdFxCtor.effectOf(a)
 
-    @inline override final def pureOf[A](a: A): Id[A] = FxCtor.IdFxCtor.pureOf(a)
+    @inline override final def pureOf[A](a: A): Id[A] = fxCtor.IdFxCtor.pureOf(a)
 
-    @inline override final val unitOf: Id[Unit] = FxCtor.IdFxCtor.unitOf
+    @inline override final val unitOf: Id[Unit] = fxCtor.IdFxCtor.unitOf
 
-    @inline override final def errorOf[A](throwable: Throwable): Id[A] = FxCtor.IdFxCtor.errorOf(throwable)
+    @inline override final def errorOf[A](throwable: Throwable): Id[A] = fxCtor.IdFxCtor.errorOf(throwable)
 
     @inline override final def mapFa[A, B](fa: Id[A])(f: A => B): Id[B] = f(fa)
