+++ effectie-monix/src/main/scala/effectie/monix/FxCtor.scala
@@ -7,8 +7,8 @@ import monix.eval.Task
 /** @author Kevin Lee
   * @since 2021-05-16
   */
-object FxCtor {
-  type FxCtor[F[_]] = effectie.core.FxCtor[F]
+object fxCtor {
+  private type FxCtor[F[_]] = effectie.core.FxCtor[F]
 
   implicit object TaskFxCtor extends FxCtor[Task] {
