+++ effectie-monix/src/test/scala/effectie/monix/syntax/errorSpec.scala
@@ -6,7 +6,7 @@ import cats.effect.IO
 import cats.syntax.all._
 import effectie.syntax.fx._
 import effectie.syntax.error._
-import effectie.monix.Fx._
+import effectie.monix.fx._
 import effectie.monix.syntax.error._
 import effectie.testing.types._
 import effectie.core.Fx
@@ -28,7 +28,6 @@ object errorSpec extends Properties {
 }
 
 object CanCatchSyntaxSpec {
-  val CanRecover: effectie.core.CanRecover.type = effectie.core.CanRecover
 
   def tests: List[Test] = taskSpecs ++ ioSpecs ++ futureSpecs ++ idSpecs
